requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.2+cu121
torchvision==0.16.2+cu121
torchaudio==2.1.2+cu121
numpy==1.24.4
ipython==8.37.0
lightning==2.2.4
h5py==3.14.0
wandb==0.20.1
pybullet==3.2.7
GitPython==3.1.44
ikfast-pybind==0.1.2
geometrout==0.1.3
torchmetrics==1.7.2
matplotlib==3.10.3
tqdm==4.67.1
urchin==0.0.29
termcolor==3.1.0
meshcat==0.3.2
moviepy==1.0.3
imageio==2.37.0
splendor-render==0.3.4
PyOpenGL==3.1.9
slack_sdk==3.35.0
yourdfpy==0.0.57
ipdb==0.13.13
torch-geometric==2.6.1
fastapi==0.115.12
numba==0.61.2
bpy==4.0.0
pyzmq==27.0.0
fasteners==0.19
urdf_parser_py==0.0.4
atob @ git+https://github.com/fishbotics/atob.git@fa15a01c66a2e825813cbac5fe0523d9b12d0203
pyg-lib==0.4.0+pt21cu121
torch-scatter==2.1.2+pt21cu121
torch-sparse==0.6.18+pt21cu121
torch-cluster==1.6.3+pt21cu121
torch-spline-conv==1.2.2+pt21cu121 